requirements.txt
google-generativeai
groq
requests
python-dotenv
